+++ Source/RethinkDb.Driver.Tests/ReQL/ChangeFeedTests.cs
@@ -4,6 +4,7 @@ using FluentAssertions;
 using Newtonsoft.Json.Linq;
 using NUnit.Framework;
 using RethinkDb.Driver.Model;
+using RethinkDb.Driver.Tests.Utils;
 
 namespace RethinkDb.Driver.Tests.ReQL
 {
@@ -20,7 +21,7 @@ namespace RethinkDb.Driver.Tests.ReQL
                 .AssertNoErrors();
 
             var changes = r.db(DbName).table(TableName)
-                .changes()[new { include_states = true }]
+                .changes()[new { include_states = false }]
                 .runChanges<JObject>(conn);
 
 
@@ -66,6 +67,65 @@ namespace RethinkDb.Driver.Tests.ReQL
 
             changes.close();
 
+            task.Result.Should().Be(3);
+        }
+
+        [Test]
+        [Explicit]
+        public void can_enumerate_though_change_feed_manually()
+        {
+            var result = r.db(DbName).table(TableName)
+                .delete()[new { return_changes = true }]
+                .runResult(conn)
+                .AssertNoErrors();
+
+            var changes = r.db(DbName).table(TableName)
+                .changes()[new { include_states = false }]
+                .runChanges<JObject>(conn);
+
+            var task = Task.Run(async () =>
+            {
+                var count = 0;
+                while (await changes.MoveNextAsync())
+                {
+                    changes.Current.Dump();
+                    count++;
+                }
+                return count;
+            });
+
+            Thread.Sleep(3000);
+
+            Task.Run(() =>
+            {
+                r.db(DbName).table(TableName)
+                    .insert(new { foo = "bar" })
+                    .run(conn);
+            });
+
+            Thread.Sleep(3000);
+
+            Task.Run(() =>
+            {
+                r.db(DbName).table(TableName)
+                    .insert(new { foo = "bar" })
+                    .run(conn);
+            });
+
+            Thread.Sleep(3000);
+
+            Task.Run(() =>
+            {
+                r.db(DbName).table(TableName)
+                    .insert(new { foo = "bar" })
+                    .run(conn);
+            });
+
+            Thread.Sleep(3000);
+
+            changes.close();
+
+            Task.WaitAll(task);
             task.Result.Should().Be(3);
         }
     }